+++ src/Method/HttpHeader.php
@@ -8,6 +8,7 @@ use JetBrains\PhpStorm\Language;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Yiisoft\Auth\AuthenticationMethodInterface;
+use Yiisoft\Auth\AuthenticatorInterface;
 use Yiisoft\Auth\IdentityInterface;
 use Yiisoft\Auth\IdentityWithTokenRepositoryInterface;
 
@@ -20,7 +21,7 @@ use function reset;
  * {@see IdentityWithTokenRepositoryInterface::findIdentityByToken()}
  * and passes the value of the `X-Api-Key` header. This implementation is used mainly for authenticating API clients.
  */
-class HttpHeader implements AuthenticationMethodInterface
+class HttpHeader implements AuthenticationMethodInterface, AuthenticatorInterface
 {
     protected string $headerName = 'X-Api-Key';
 
@@ -43,6 +44,10 @@ class HttpHeader implements AuthenticationMethodInterface
         return null;
     }
 
+    /**
+     * @deprecated No-op kept only for compatibility with the deprecated {@see AuthenticationMethodInterface}.
+     * HTTP header authentication does not need a challenge.
+     */
     public function challenge(ResponseInterface $response): ResponseInterface
     {
         return $response;